Remove npm Dockerfile and migrate its functionality to php Dockerfile
This commit removes the separate npm Dockerfile and incorporates its relevant functionalities directly into the php Dockerfile. In the process, it updates the entry points for npm and ncu in the php Dockerfile. Furthermore, the reference to the now non-existing npm Dockerfile are replaced with the php Dockerfile in docker-compose files. Also, the PostgreSQL container name has been changed from 'postgres' to 'db'.