Update Github workflows configuration
Modified the test.yml and deploy.yml in the Github workflows to include the "master" branch in the pull request triggers for test.yml and remove the "master" branch from the pull request triggers in deploy.yml. This change ensures that code tests are run for pull requests on both development and master branches, while deployment is only triggered on push to master.