Refactor tests and update telegram connection and subscription features
This commit refactors various tests of features like password resetting, password confirmation and email verification. It also updates the telegram connect functionality to require authorization and adds validation for the 'payment' query and 'newSubscription' request in Dashboard and Subscription controllers respectively. Lastly, a new method for subscriptions was added in the LemonSqueezyProduct enum.