Refactor Telegram service and notifications behavior
The main update in this commit involves the refactoring of the Telegram Service and the modification of the related notifications behavior. Additionally, the "ReplyToUnknownCommand" and "UnknownMessageReceived" classes no longer have the 'to' method. Also, "TelegramController" now redirects to 'login' instead of 'home' if no 'tgid' query parameter is present. Conditions in the 'console.php', 'UserObserver.php', and 'TelegramService.php' now use 'app()->isProduction' instead of comparing 'app.env' with 'production' directly.