Refactored Telegram notifications (#12)
* Enable telegram notifications on user account connection

The `TelegramController` now updates the user's `send_tg_notifications` field to `true` along with the `telegram_id` during the user update process. This means when the user connects their account to Telegram, they will automatically opt-in to receiving Telegram notifications.

* Refactor Telegram service and notifications behavior

The main update in this commit involves the refactoring of the Telegram Service and the modification of the related notifications behavior. Additionally, the "ReplyToUnknownCommand" and "UnknownMessageReceived" classes no longer have the 'to' method. Also, "TelegramController" now redirects to 'login' instead of 'home' if no 'tgid' query parameter is present. Conditions in the 'console.php', 'UserObserver.php', and 'TelegramService.php' now use 'app()->isProduction' instead of comparing 'app.env' with 'production' directly.