Cypress end-to-end tests (#14)
* Add script to check and restart Docker containers

A bash script to check if specified Docker containers are running, and restart the Docker Compose setup if any container is not running has been added. Also, the deployment action has been updated to stop and start the cron service during deployment because of the script.

* Update package dependencies and add Cypress to docker-compose

This commit updates a variety of package dependencies, most notably adding Cypress. The Prettier and prettier-plugin-tailwindcss versions have also been updated. A chunk of new dependencies related to Cypress and other updates have been added. Additionally, a new service 'Cypress' has been introduced for end-to-end testing.

* Add Cypress for end-to-end testing

This commit introduces Cypress as a tool for end-to-end testing. It adds various tests for checking user authentication, CRUD operations, and static pages interactions. The build-and-test workflow has been adjusted to include Cypress tests. The application code also contains some slight changes to accommodate testing needs. This integration will enhance the quality and reliability of the application.

* Update .gitignore, docker-compose-ci.yml, routes.json, UserService & action.yml

.gitignore file has been updated to ignore cypress screenshots.
Modified UserService to return fake data for Lemon Squeezy links if LS key is not available and app is not in production.

Lastly, the GitHub Actions workflow file has been updated to replace SQLite as the default DB connection for Cypress tests.