Add page view tracking and syncing
This commit adds page view tracking and syncing to the application. Introduced `LogViews` middleware that records the views for each page in Redis. Also added a daily command to sync these views from Redis to the database. Additionally, included redis in the `docker-compose` files and updated appropriate middleware in the routes files.