MyEvents download button styles update
Improve color contrast for better accessibility

This commit improves the color contrast on Dashboard and MyEvents pages to aid people with visual impairments by making the texts easier to read. Further adjusts the border and text colors in light and dark themes to meet the contrast standards for better usability and accessibility.